+++ CHANGELOG.md
@@ -10,6 +10,11 @@
 * Breaking Changes
   * CardPayments
     * Remove `Vault` data class originally meant for Vault with Purchase flows (it's no longer needed)
+  * CorePayments
+    * Remove `STAGING` constant from `Environment` enum
+  * FraudDetection
+    * Update `PayPalDataCollector` constructor to require a configuration instead of an environment
+    * Remove `PayPalDataCollectorEnvironment` enum
 * PayPalNativePayments
   * Bump `PayPal Native Checkout` to `1.1.0`
 
@@ -23,9 +28,7 @@
   * Make `Address` implement `Parcelable`
 * PayPalNativePayments
   * Bump `PayPal Native Checkout` to `1.0.0`
-* Breaking Changes
-  * CorePayments
-    * Remove `STAGING` constant from `Environment` enum
+
 
 ## 0.0.10 (2023-06-23)
 * Breaking Changes